// api/models/delivery-address.ts
/* tslint:disable */
/* eslint-disable */
/**
 * PardoMart Node API
 * API documentation for the PardoMart Node.js application.
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: support@pardomart.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface DeliveryAddress {
    'id'?: string;
    'userId'?: string;
    'addressLine1'?: string;
    'addressLine2'?: string | null;
    'city'?: string;
    'state'?: string | null;
    'postalCode'?: string | null;
    'country'?: string;
    'isDefault'?: boolean;
    'latitude'?: number | null;
    'longitude'?: number | null;
    'createdAt'?: string;
    'updatedAt'?: string;
    'label'?: string | null;
}